third_party/blink/web_tests/external/wpt/fetch/range/sw.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Accept-Encoding should not appear in a service worker
  assert_equals: Accept-Encoding should not be set for media expected (object) null but got (string) "identity;q=1, *;q=0"
Harness: the test ran to completion.
